+++ elfpy/utils/json.py
@@ -1,0 +1,38 @@
+"""Extensons to json encoding."""
+import json
+
+import numpy as np
+from fixedpointmath import FixedPoint
+from hexbytes import HexBytes
+from numpy.random._generator import Generator as NumpyGenerator
+from web3.datastructures import AttributeDict, MutableAttributeDict
+
+
+class ExtendedJSONEncoder(json.JSONEncoder):
+    r"""Custom encoder for JSON string dumps"""
+    # pylint: disable=too-many-return-statements
+
+    def default(self, o):
+        r"""Override default behavior"""
+        if isinstance(o, set):
+            return list(o)
+        if isinstance(o, HexBytes):
+            return o.hex()
+        if isinstance(o, (AttributeDict, MutableAttributeDict)):
+            return dict(o)
+        if isinstance(o, np.integer):
+            return int(o)
+        if isinstance(o, np.floating):
+            return float(o)
+        if isinstance(o, np.ndarray):
+            return o.tolist()
+        if isinstance(o, FixedPoint):
+            return str(o)
+        if isinstance(o, NumpyGenerator):
+            return "NumpyGenerator"
+        try:
+            return o.__dict__
+        except AttributeError:
+            pass
+        # Let the base class default method raise the TypeError
+        return json.JSONEncoder.default(self, o)
