+++ elfpy/utils/outputs.py
@@ -1,16 +1,11 @@
 """Helper functions for delivering simulation outputs"""
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING, Any
 
 import matplotlib.pyplot as plt
 import numpy as np
-from fixedpointmath import FixedPoint
-from hexbytes import HexBytes
 from matplotlib import gridspec
-from numpy.random._generator import Generator as NumpyGenerator
-from web3.datastructures import AttributeDict, MutableAttributeDict
 
 if TYPE_CHECKING:
     import pandas as pd
@@ -404,33 +399,3 @@ def annotate(axis_handle, text, major_offset, minor_offset, val):
     annotation_handle.set_bbox(
         {"facecolor": "white", "edgecolor": "black", "alpha": val["alpha"], "linewidth": 0, "boxstyle": "round,pad=0.1"}
     )
-
-
-class ExtendedJSONEncoder(json.JSONEncoder):
-    r"""Custom encoder for JSON string dumps"""
-    # pylint: disable=too-many-return-statements
-
-    def default(self, o):
-        r"""Override default behavior"""
-        if isinstance(o, set):
-            return list(o)
-        if isinstance(o, HexBytes):
-            return o.hex()
-        if isinstance(o, (AttributeDict, MutableAttributeDict)):
-            return dict(o)
-        if isinstance(o, np.integer):
-            return int(o)
-        if isinstance(o, np.floating):
-            return float(o)
-        if isinstance(o, np.ndarray):
-            return o.tolist()
-        if isinstance(o, FixedPoint):
-            return str(o)
-        if isinstance(o, NumpyGenerator):
-            return "NumpyGenerator"
-        try:
-            return o.__dict__
-        except AttributeError:
-            pass
-        # Let the base class default method raise the TypeError
-        return json.JSONEncoder.default(self, o)
